busy rearranging into components
